[etl] initial commit
I had lots of troubles, to build this hello world. The main problem was
that I was using BOOST-1.61.0, and the testing code fails.
I was not able to understand the error message of the compiler since it
was this sort of template errors.

I've installed BOOST-1.55.0, and it worked like a champ. The problem was
not mine.